+++ service/scheduler_handler.cpp
@@ -62,7 +62,7 @@ boost::posix_time::seconds parse_interval(const std::string &str) {
 void scheduler::add_task(schedule_metadata::task_source source, std::string interval, const std::string info) {
   const auto metrics_interval = parse_interval(interval);
   if (source == schedule_metadata::METRICS) {
-    metrics_interval_ = metrics_interval.total_seconds();
+    metrics_interval_ = static_cast<unsigned int>(metrics_interval.total_seconds());
   }
   unsigned int id = tasks.add_task("internal", metrics_interval, 0.5);
   schedule_metadata data;
